# packages/backend/openapi.yaml
openapi: 3.0.3
info:
  title: ID PASS DataCollect Backend API
  description: |
    REST API for the ID PASS DataCollect backend server. The server is the central
    sync hub for offline-first DataCollect clients (mobile, web, admin UI) and
    coordinates external integrations via the adapter registry.

    ## Authentication

    Most endpoints require a JWT access token in the `Authorization` header:

    ```
    Authorization: Bearer <token>
    ```

    Access tokens are issued by `POST /api/users/login` and expire after one hour.
    Use `POST /api/users/refresh` to get a new access token (and rotated refresh token)
    without re-prompting for credentials — field agents use this to stay signed in
    while offline. Self-service endpoints (`/api/auth/**`) issue separate scoped
    tokens via OTP, national ID, or OIDC exchange.

    ## Multi-tenant (multi-program) support

    All tenant-scoped endpoints accept a `configId` query parameter (or field in
    the request body). Each configuration has its own entity forms, entity data,
    auth providers, and external-sync settings. Non-admin users can only access
    configurations listed in their `tenantIds` claim — the `validateTenantAccess`
    middleware enforces this on every tenant-scoped route.

    ## Synchronization

    The sync endpoints support bidirectional synchronization between clients and
    server, as well as external system integration. Internal sync uses cursor-based
...
    persisted to `sync_events`.
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin and user JWT issued by `/api/users/login` or `/api/users/refresh`
    SelfServiceAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Self-service JWT issued by `/api/auth/otp/verify`, `/api/auth/id/verify`, or `/api/auth/oidc/exchange`
...
  schemas:
...
    OpenSppField:
      type: object
      properties:
        name: { type: string }
        type:
          type: string
          enum: [text, date, relation, selection]
        label: { type: string }
        required: { type: boolean }
        options:
          type: array
          items:
            type: object
            properties:
              id: {}
              label: { type: string }
...
paths:
...
  /api/openspp-fields/parse-file:
    post:
      tags: [OpenSPP Fields]
      summary: Parse OpenSPP fields from uploaded JSON file
      description: Upload a JSON file; the server infers field types and returns the schema. Max 10 MB.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [payload]
              properties:
                payload:
                  type: string
                  format: binary
      responses:
        '200':
          description: Parsed fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  fields:
                    type: array
                    items: { $ref: '#/components/schemas/OpenSppField' }
        '400':
          description: Invalid JSON or missing file
  /api/openspp-fields/parse:
    post:
      tags: [OpenSPP Fields]
      summary: Parse OpenSPP fields from JSON payload
      description: Inline JSON variant of `/parse-file`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Parsed fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  fields:
                    type: array
                    items: { $ref: '#/components/schemas/OpenSppField' }
  /api/openspp-fields/fetch:
    post:
      tags: [OpenSPP Fields]
      summary: Fetch fields from OpenSPP V1 (Odoo)
      description: |
        Authenticates to Odoo via JSON-RPC and returns the field metadata for the given model.
        Blocks requests to private IP ranges and cloud metadata endpoints (SSRF mitigation).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, database, username, password]
              properties:
                url: { type: string, format: uri }
                database: { type: string }
                username: { type: string }
                password: { type: string }
                model: { type: string, default: res.partner }
                fields:
                  type: array
                  items: { type: string }
                attributes:
                  type: array
                  items: { type: string }
      responses:
        '200':
          description: Fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  fields:
                    type: array
                    items: { $ref: '#/components/schemas/OpenSppField' }
        '400':
          description: Missing credentials or URL points at a private/internal range
        '500':
          description: Failed to fetch from Odoo
...
security:
  - BearerAuth: []